Move worthwhile JDK 26 enum tests into rewrite-java-tck
The annotation and lambda tests duplicated existing TCK coverage
(AnnotationTest.annotationWithDefaultArgument /
LambdaTest.untypedLambdaParameter) — those already exercise the JDK 26
fix once rewrite-java-next:compatibilityTest runs on JDK 26.

The two enum cases with non-literal arguments (method call, string
concat) are genuinely new ground, so move them to EnumTest in the TCK
where they run against every parser module.
