--- rewrite-java-next/src/test/java/org/openrewrite/java/Java26EnumConstructorTest.java
@@ -1,117 +1,0 @@
-/*
- * Copyright 2026 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.java;
-
-import org.junit.jupiter.api.Test;
-import org.openrewrite.Issue;
-import org.openrewrite.test.RewriteTest;
-
-import static org.openrewrite.java.Assertions.java;
-
-class Java26EnumConstructorTest implements RewriteTest {
-
-    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
-    @Test
-    void simpleEnumConstantWithStringArg() {
-        rewriteRun(
-          java(
-            """
-              enum Action {
-                  COPY_TO("Copy to"),
-                  COPY_MORE("Copy");
-
-                  private final String text;
-
-                  Action(String text) {
-                      this.text = text;
-                  }
-              }
-              """
-          )
-        );
-    }
-
-    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
-    @Test
-    void enumConstantWithMethodCallArg() {
-        rewriteRun(
-          java(
-            """
-              enum Action {
-                  COPY_TO(label("Copy to")),
-                  COPY_MORE(label("Copy"));
-
-                  private final String text;
-
-                  Action(String text) {
-                      this.text = text;
-                  }
-
-                  static String label(String s) {
-                      return s;
-                  }
-              }
-              """
-          )
-        );
-    }
-
-    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
-    @Test
-    void enumConstantWithStringConcat() {
-        rewriteRun(
-          java(
-            """
-              enum Action {
-                  COPY_TO("Copy to"),
-                  COPY_MORE("Copy" + "...");
-
-                  private final String text;
-
-                  Action(String text) {
-                      this.text = text;
-                  }
-              }
-              """
-          )
-        );
-    }
-
-    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
-    @Test
-    void enumConstantWithMethodCallAndConcat() {
-        rewriteRun(
-          java(
-            """
-              enum Action {
-                  COPY_TO(label("Copy to")),
-                  COPY_MORE(label("Copy") + "...");
-
-                  private final String text;
-
-                  Action(String text) {
-                      this.text = text;
-                  }
-
-                  static String label(String s) {
-                      return s;
-                  }
-              }
-              """
-          )
-        );
-    }
-}
