+++ rewrite-java-tck/src/main/java/org/openrewrite/java/tree/EnumTest.java
@@ -135,6 +135,52 @@ class EnumTest implements RewriteTest {
         );
     }
 
+    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
+    @Test
+    void enumConstantWithMethodCallArgument() {
+        rewriteRun(
+          java(
+            """
+              enum Action {
+                  COPY_TO(label("Copy to")),
+                  COPY_MORE(label("Copy"));
+
+                  private final String text;
+
+                  Action(String text) {
+                      this.text = text;
+                  }
+
+                  static String label(String s) {
+                      return s;
+                  }
+              }
+              """
+          )
+        );
+    }
+
+    @Issue("https://github.com/openrewrite/rewrite/issues/7554")
+    @Test
+    void enumConstantWithStringConcatArgument() {
+        rewriteRun(
+          java(
+            """
+              enum Action {
+                  COPY_TO("Copy to"),
+                  COPY_MORE("Copy" + "...");
+
+                  private final String text;
+
+                  Action(String text) {
+                      this.text = text;
+                  }
+              }
+              """
+          )
+        );
+    }
+
     @Test
     void enumWithoutParameters() {
         rewriteRun(
